requirements/requirements.txt
mongoengine==0.17.0
pymongo==3.8.0
celery==4.3.0
requests==2.22.0
requests-toolbelt==0.9.1
pika==1.0.1
fabric3==1.14.post1
tldextract==2.2.1
dnspython==1.16.0
gevent==1.4.0
argparse==1.4.0
IPy==1.0
python-nmap==0.6.1
geoip2==2.9.0
beautifulsoup4==4.8.0
shodan==1.14.0
sentry-sdk==0.10.2
lxml==4.4.1
bs4
gevent==1.4.0
progressbar==2.5
python-nmap==0.6.1
dnspython==1.16.0
IPy==1.0
geoip2==2.9.0
click=7.0






